Rename validator services
