+++ src/Thenewboston/Validator/Api/IAccountsService.cs
@@ -6,7 +6,7 @@ using Thenewboston.Validator.Models;
 
 namespace Thenewboston.Validator.Api
 {
-    public interface IValidatorApiClient
+    public interface IAccountsService
     {
         Task<IEnumerable<ValidatorAccount>> GetAccountsAsync();
         Task<ValidatorAccountBalance> GetAccountBalanceAsync(string accountNumber);
